Always include column names for initial import.
When CKE5 profiles are included with ydeploy, the generated initial dump
omits most columns (including `id`) and mixes the order of the columns
resulting in a dump that's not importable. This behaviour is observable
with mariadb-dump > 10.11.0. The import aborts with a message of a
duplicate PRIMARY KEY '0'.

Using the `-c` parameter dumps with the column names included, allowing
to import the generated dump.